feature: AppDelegate bootstrap code.
This adds AppDelegate bootstrap code which generates a swift IR which
exposes the AppDelegate functions through CCC stubs. These are binded to
and linked which allow for you to at least boot. Further research will
be needed to be done on how to connect ViewControllers since those need
to be subclassed. Likely the Objective-C API will be the best bet or
rather a `VSLViewController` can be subclassed which basically exposes
100% of the things a `ViewController` can do and that can be used for
interaction. This does require the addition of external initializers.

Signed-off-by: Vihan <contact@vihan.org>